+++ bin/run_unit_tests.sh
@@ -16,15 +16,12 @@ do
     esac
 done
 
-USER=$(stat -c "%u" "$0")
-GROUP=$(stat -c "%g" "$0")
-
 docker run \
     --rm \
     --security-opt "${DEV_ENV_SECURITY_OPT:-seccomp:unconfined}" \
     --volume "$DEV_ENV_SOURCE_CODE:/app" \
     "$DEV_ENV_DOCKER_IMAGE" \
-    run_unit_tests.sh "$DEV_ENV_PACKAGE" "$USER" "$GROUP" "$@"
+    run_unit_tests.sh "$DEV_ENV_PACKAGE" "$@"
 
 sed -i -e "s|/app/|$DEV_ENV_SOURCE_CODE/|g" "$DEV_ENV_SOURCE_CODE/.coverage"