+++ ubuntu/xenial/run_unit_tests.sh
@@ -2,18 +2,12 @@
 
 set -e
 
-if [ $# -lt 3 ]; then
-    echo "usage: $(basename "$0") <package> <user> <group> [<dir1> <dir2> ... <dirN>]" >&2
+if [ $# -lt 1 ]; then
+    echo "usage: $(basename "$0") <package> [<dir1> <dir2> ... <dirN>]" >&2
     exit 1
 fi
 
 PACKAGE=${1:-}
-shift
-
-USER=${1:-}
-shift
-
-GROUP=${1:-}
 shift
 
 nosetests \
@@ -22,7 +16,5 @@ nosetests \
     --cover-branches \
     "--cover-package=$PACKAGE" \
     "$@"
-
-chown "$USER.$GROUP" "/app/.coverage"
 
 exit 0