+++ CHANGELOG.md
@@ -12,7 +12,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 
-- Nothing
+- ```run_unit_tests.sh``` no longer chown's ```/app/.coverage```
 
 ### Removed